extract label-editor
